requirements.txt
streamlit
tensorflow
numpy
Pillow
matplotlib
scikit-learn
plotly